Use selected version as description message when generating completions.
Another alternative was to set the description to an empty string.

This fixes a silent bug that was occurring only when the user activated
a node version other than the lts or latest and tried to use the node
binary inside the same directory where the .nvmrc/.fnmrc file was located.