Replace usage of fish_user_paths with symlinks
fish_user_paths has the unfortunate side effect that if it points
to a path somewhere in the user's home directory, the path will be
set to the full path (i.e. /home/username/ instead of ~/) inside
the fish_variables file.

This is unfortunate because a user that wish to share the
fish_variables file between multiple computers, to get a unified
shell setup, it's not possible since the path will point to a home
directory that might not be the same between the host (since the
user possibly has different usernames on different hosts).

This aims to rememdy this by using symlinks instead of
the fish_user_paths variable.